Conflicts:
	lib/MediaWords/Controller/Admin/CM.pm

+++ install_scripts/install_mediacloud_system_package_dependencies.sh
@@ -103,6 +103,8 @@ else
 
     # assume Ubuntu
 
+    source /etc/lsb-release
+
     # Add 10gen repository
     APT_SOURCES_MONGODB="/etc/apt/sources.list.d/mongodb.list"
     if [ ! -f "$APT_SOURCES_MONGODB" ]; then
@@ -128,14 +130,18 @@ else
         }
     done
 
-    # Apt's version of Gearman is too old
-    sudo apt-get -y remove gearman gearman-job-server gearman-tools \
-        libgearman-dbg libgearman-dev libgearman-doc libgearman6
-
     # Install Gearman from PPA repository
     sudo apt-get -y install python-software-properties
-    sudo add-apt-repository -y ppa:gearman-developers/ppa
-    sudo apt-get -y update
+
+    if [[ "$DISTRIB_RELEASE" < "14.04" ]]; then
+         # Apt's version of Gearman is too old
+	sudo apt-get -y remove gearman gearman-job-server gearman-tools \
+        libgearman-dbg libgearman-dev libgearman-doc libgearman6
+
+	sudo add-apt-repository -y ppa:gearman-developers/ppa
+	sudo apt-get -y update
+    fi
+
     sudo apt-get -y install gearman-job-server gearman-tools libgearman-dev
 
     # Install the rest of the packages
